hoàn thiện 80% admin

+++ src/main/java/com/poly/bezbe/dto/response/product/ProductBriefDTO.java
@@ -9,5 +9,5 @@ import lombok.Data;
 public class ProductBriefDTO {
     private Long id;
     private String name;
-    private Integer variantCount;
+    private Long variantCount;
 }
